requirements.txt
pyahocorasick==2.0.0
fastapi==0.115.14
httpx==0.28.1
uvicorn==0.35.0
mysql-connector-python==9.3.0
tqdm==4.67.1
